// .PRD/erp-refactor/v3/types/quotation.ts
export const QuotationSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  naming_series: z.enum(['SAL-QTN-.YYYY.-']),
  quotation_to: z.string().default('Customer'),
  party_name: z.string().optional(),
  customer_name: z.string().optional(),
  transaction_date: z.string().default('Today'),
  valid_till: z.string().optional(),
  order_type: z.enum(['Sales', 'Maintenance', 'Shopping Cart']).default('Sales'),
  company: z.string(),
  has_unit_price_items: z.boolean().optional().default(false),
  amended_from: z.string().optional(),
  currency: z.string(),
  conversion_rate: z.number(),
  selling_price_list: z.string(),
  price_list_currency: z.string(),
  plc_conversion_rate: z.number(),
  ignore_pricing_rule: z.boolean().optional().default(false),
  scan_barcode: z.string().optional(),
  last_scanned_warehouse: z.string().optional(),
  items: z.array(z.unknown()),
  total_qty: z.number().optional(),
  total_net_weight: z.number().optional(),
  base_total: z.number().optional(),
  base_net_total: z.number().optional(),
  total: z.number().optional(),
  net_total: z.number().optional(),
  tax_category: z.string().optional(),
  taxes_and_charges: z.string().optional(),
  shipping_rule: z.string().optional(),
  incoterm: z.string().optional(),
  named_place: z.string().optional(),
  taxes: z.array(z.unknown()).optional(),
  base_total_taxes_and_charges: z.number().optional(),
  total_taxes_and_charges: z.number().optional(),
  grand_total: z.number().optional(),
  in_words: z.string().max(240).optional(),
  rounded_total: z.number().optional(),
  rounding_adjustment: z.number().optional(),
  disable_rounded_total: z.boolean().optional().default(false),
  base_grand_total: z.number().optional(),
  base_in_words: z.string().max(240).optional(),
  base_rounded_total: z.number().optional(),
  base_rounding_adjustment: z.number().optional(),
  apply_discount_on: z.enum(['Grand Total', 'Net Total']).optional().default('Grand Total'),
  base_discount_amount: z.number().optional(),
  coupon_code: z.string().optional(),
  additional_discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  referral_sales_partner: z.string().optional(),
  other_charges_calculation: z.string().optional(),
  item_wise_tax_details: z.array(z.unknown()).optional(),
  packed_items: z.array(z.unknown()).optional(),
  pricing_rules: z.array(z.unknown()).optional(),
  customer_address: z.string().optional(),
  address_display: z.string().optional(),
  contact_person: z.string().optional(),
  contact_display: z.string().optional(),
  contact_mobile: z.string().optional(),
  contact_email: z.string().email().optional(),
  shipping_address_name: z.string().optional(),
  shipping_address: z.string().optional(),
  company_address: z.string().optional(),
  company_address_display: z.string().optional(),
  company_contact_person: z.string().optional(),
  payment_terms_template: z.string().optional(),
  payment_schedule: z.array(z.unknown()).optional(),
  tc_name: z.string().optional(),
  terms: z.string().optional(),
  auto_repeat: z.string().optional(),
  letter_head: z.string().optional(),
  group_same_items: z.boolean().optional().default(false),
  select_print_heading: z.string().optional(),
  language: z.string().optional(),
  lost_reasons: z.array(z.unknown()).optional(),
  competitors: z.array(z.unknown()).optional(),
  order_lost_reason: z.string().optional(),
  status: z.enum(['Draft', 'Open', 'Replied', 'Partially Ordered', 'Ordered', 'Lost', 'Cancelled', 'Expired']).default('Draft'),
  customer_group: z.string().optional(),
  territory: z.string().optional(),
  utm_source: z.string().optional(),
  utm_campaign: z.string().optional(),
  utm_medium: z.string().optional(),
  utm_content: z.string().optional(),
  opportunity: z.string().optional(),
  supplier_quotation: z.string().optional(),
  enq_det: z.string().optional(),
});

export type Quotation = z.infer<typeof QuotationSchema>;

export const QuotationInsertSchema = QuotationSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });